backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
twilio==9.3.6
python-dotenv==1.0.1
